+++ src/wskr/__init__.py
@@ -1,7 +1,7 @@
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 
-from . import config as _config
+from .core import config as _config
 
 # Make *absolutely sure* we never try to invoke TeX during draw or savefig,
 # which would hang under test timeouts.
@@ -13,7 +13,7 @@ def init(*, style: str | None = None, **config_overrides: object) -> None:
     if config_overrides:
         _config.configure(**config_overrides)
     if style == "auto-dark":
-        from .mpl.utils import detect_dark_mode  # noqa: PLC0415
+        from .render.matplotlib.utils import detect_dark_mode  # noqa: PLC0415
 
         if detect_dark_mode():
             plt.style.use("dark_background")