+++ src/wskr/render/matplotlib/kitty.py
@@ -1,0 +1,26 @@
+import sys
+
+from matplotlib import _api, interactive
+from matplotlib.backend_bases import _Backend
+from matplotlib.backends.backend_agg import FigureCanvasAgg
+
+from wskr.kitty.transport import KittyTransport
+from wskr.mpl.base import BaseFigureManager, TerminalBackend
+
+if sys.flags.interactive:
+    interactive(b=True)
+
+
+class KittyFigureManager(BaseFigureManager):
+    def __init__(self, canvas: FigureCanvasAgg, num: int = 1):
+        super().__init__(canvas, num, KittyTransport)
+
+
+class KittyFigureCanvas(FigureCanvasAgg):
+    manager_class = _api.classproperty(lambda _: KittyFigureManager)
+
+
+@_Backend.export
+class _BackendKittyAgg(TerminalBackend):
+    FigureCanvas = KittyFigureCanvas
+    FigureManager = KittyFigureManager